Feat: Allow `label` in `MenuItem` to be a `Widget`
The `label` property of the `MenuItem` class has been changed from a `String` to a `Widget`. This provides more flexibility, allowing for custom text styling or other widgets to be used as a menu item's label.

The example has been updated to wrap all `String` labels with a `Text` widget to reflect this change.